requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
feedparser>=6.0.0
python-dotenv>=0.19.0
requests>=2.26.0
lark-parser>=0.12.0